+++ internal/database/database.go
@@ -1,0 +1,13 @@
+package database
+
+import (
+
+)
+
+type Database interface {
+	Connect(conn string) error
+	Disconnect() error
+
+	ProcessQueue() error
+	GetQueueSize() int
+}