+++ internal/static/static.go
@@ -1,0 +1,13 @@
+package static
+
+import (
+
+)
+
+const (
+	APIVersion = "/api/v2"
+)
+
+var (
+	Version = "canary"
+)